+++ src/app/api/reviews/create/route.ts
@@ -22,7 +22,7 @@ export async function POST(req: NextRequest) {
     }
 
     const year = parseInt(yearStr as string);
-    const rating = parseInt(ratingStr as string);
+    const rating = parseFloat(ratingStr as string);
 
     if (isNaN(year) || String(yearStr).length !== 4) {
       return NextResponse.json({ message: 'invalid year format' }, { status: 400 });
